fix(webhook): rework tests structure
