+++ connector-runtime/connector-runtime-spring/src/test/java/io/camunda/connector/runtime/inbound/webhook/WebhookExecutablesTest.java
@@ -16,46 +16,29 @@
  */
 package io.camunda.connector.runtime.inbound.webhook;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.*;
 
-import io.camunda.connector.runtime.core.inbound.InboundConnectorReportingContext;
+import io.camunda.connector.api.inbound.Health;
 import io.camunda.connector.runtime.inbound.executable.RegisteredExecutable;
-import java.util.LinkedList;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-public class WebhookExecutablesTest {
-  private RegisteredExecutable.Activated connector1;
-  private RegisteredExecutable.Activated connector2;
-  private RegisteredExecutable.Activated connector3;
-  private InboundConnectorReportingContext context1;
-  private InboundConnectorReportingContext context2;
-  private InboundConnectorReportingContext context3;
-  private static final String CONTEXT = "testContext";
-  private static final String CONTEXT2 = "otherContext";
-
-  @BeforeEach
-  void setUp() {
-    context1 = mock(InboundConnectorReportingContext.class);
-    context2 = mock(InboundConnectorReportingContext.class);
-    context3 = mock(InboundConnectorReportingContext.class);
-    connector1 = new RegisteredExecutable.Activated(null, context1);
-    connector2 = new RegisteredExecutable.Activated(null, context2);
-    connector3 = new RegisteredExecutable.Activated(null, context3);
-    when(context1.getLogs()).thenReturn(new LinkedList<>());
-    when(context2.getLogs()).thenReturn(new LinkedList<>());
-    when(context3.getLogs()).thenReturn(new LinkedList<>());
-  }
+public class WebhookExecutablesTest extends WebhookTestsBase {
 
   @Test
   void testAddConnectorToQueue() {
-    WebhookExecutables executables = new WebhookExecutables(connector1, CONTEXT);
-    executables.markAsDownAndAdd(connector2);
+    RegisteredExecutable.Activated activeExecutable = buildConnector("processA", 1, "myPath");
+    WebhookExecutables executables = new WebhookExecutables(activeExecutable, "myPath");
+    executables.markAsDownAndAdd(buildConnector("processB", 1, "myPath"));
 
-    assertEquals(2, executables.getExecutables().size());
-    assertTrue(executables.getExecutables().contains(connector1));
-    assertTrue(executables.getExecutables().contains(connector2));
+    assertThat(executables.getActiveWebhook()).isEqualTo(activeExecutable);
+    assertThat(executables.getDownExecutables()).hasSize(1);
+    var downExecutable = executables.getDownExecutables().getFirst();
+    assertThat(downExecutable.context().getHealth())
+        .isEqualTo(
+            Health.down(
+                new IllegalStateException(
+                    "Context: myPath already in use by: process processA(testElement)")));
   }
 
   @Test
